Initial refactor. Tests need updating
